moves QueryPatternUtils into 'se.liu.ida.hefquin.base.query.utils' package
