+++ hefquin-base/src/main/java/se/liu/ida/hefquin/base/query/impl/BGPImpl.java
@@ -18,6 +18,7 @@ import se.liu.ida.hefquin.base.query.BGP;
 import se.liu.ida.hefquin.base.query.SPARQLGraphPattern;
 import se.liu.ida.hefquin.base.query.TriplePattern;
 import se.liu.ida.hefquin.base.query.VariableByBlankNodeSubstitutionException;
+import se.liu.ida.hefquin.base.query.utils.QueryPatternUtils;
 
 public class BGPImpl implements BGP
 {
