+++ hefquin-base/src/main/java/se/liu/ida/hefquin/base/query/impl/TriplePatternImpl.java
@@ -12,6 +12,7 @@ import se.liu.ida.hefquin.base.query.BGP;
 import se.liu.ida.hefquin.base.query.SPARQLGraphPattern;
 import se.liu.ida.hefquin.base.query.TriplePattern;
 import se.liu.ida.hefquin.base.query.VariableByBlankNodeSubstitutionException;
+import se.liu.ida.hefquin.base.query.utils.QueryPatternUtils;
 import se.liu.ida.hefquin.jenaext.sparql.syntax.ElementUtils;
 
 import java.util.Collections;
